Gives functional TSCH for Atmel on lastest Contiki branch. Tuning can be better. Keeps 10ms Orchestra unicast in sync.

+++ cpu/avr/rtimer-arch.h
@@ -62,7 +62,7 @@
  * Round to nearest integer rather than truncate. */
 #define US_TO_RTIMERTICKS(US)  ((US) >= 0 ?                        \
                                (((int32_t)(US) * (RTIMER_ARCH_SECOND) + 500000) / 1000000L) :      \
-                               ((int32_t)(US) * (RTIMER_ARCH_SECOND) - 500000) / 1000000L)
+				((int32_t)(US) * (int32_t)(RTIMER_ARCH_SECOND) - 500000) / 1000000L)
 
 #define RTIMERTICKS_TO_US(T)   ((T) >= 0 ?                     \
                                (((int32_t)(T) * 1000000L + ((RTIMER_ARCH_SECOND) / 2)) / (RTIMER_ARCH_SECOND)) : \
